auth_system/authentication.py
import jwt
from django.conf import settings
from django.contrib.auth import get_user_model
from rest_framework.authentication import BaseAuthentication
from rest_framework.exceptions import AuthenticationFailed

User = get_user_model()


class JWTAuthentication(BaseAuthentication):

    def authenticate(self, request):
        header = request.headers.get("Authorization")
        if not header or not header.startswith("Bearer "):
            return None
        token = header.split()[1]

        try:
            payload = jwt.decode(
                token,
                settings.JWT_SECRET_KEY,
                algorithms=["HS256"],
            )
        except Exception:
            raise AuthenticationFailed("Время жизни JWT истекло или неправильный JWT") from None

        user_id = payload.get("user_id")
        if not user_id:
            raise AuthenticationFailed("Неправильный JWT")

        try:
            user = User.objects.get(id=user_id)
        except User.DoesNotExist:
            raise AuthenticationFailed("Пользователь не определен") from None

        return (user, None)
